refactor: filter by formType after fetching
